fix: eliminate all compilation and clippy warnings across all projects
- Fixed 14 compilation warnings (unused imports, unnecessary parens,
  unused variables) in kernel
- Fixed ~450 clippy style warnings across kernel, drivers, NXL, LSP,
  bootloader, and userbin
- Removed redundant #[test] blocks from page_cache.rs
- Added custom_test_frameworks noop runner for no_std targets
- Added nightly rust-toolchain.toml to all no_std projects
- Added test=false to all no_std Cargo.toml targets
- Added .vscode/settings.json for rust-analyzer config
- Stress test: 300/300 PASS

+++ libneodos-nxl/src/mem.rs
@@ -32,7 +32,7 @@ pub extern "C" fn nxl_sbrk(increment: i64) -> i64 {
     let current = nxl_sys_brk(0);
     if current < 0 { return current; }
     if increment == 0 { return current; }
-    let new = (current as i64).checked_add(increment).unwrap_or(-1);
+    let new = current.checked_add(increment).unwrap_or(-1);
     if new < 0 { return EINVAL; }
     let new = new as u64;
     let result = nxl_sys_brk(new);
